Group all tests in tests folder
