+++ package.json
@@ -35,12 +35,12 @@
     "postinstall": "opencollective-postinstall || true",
     "test": "npm run test:unit && npm run test:e2e",
     "test:ci": "npm run test:unit && npm run test:e2e:ci",
-    "test:unit": "vitest run --config ./vitest/vitest.config.ts",
-    "test:unit:watch": "vitest --config ./vitest/vitest.config.ts",
-    "test:e2e": "npx playwright test",
+    "test:unit": "vitest run --config ./tests/config/vitest.config.ts",
+    "test:unit:watch": "vitest --config ./tests/config/vitest.config.ts",
+    "test:e2e": "npx playwright test --config ./tests/config/playwright.config.ts",
     "test:e2e:dev": "npx playwright test --ui",
     "test:e2e:instrument": "nyc instrument --compact=false dist tests/fixtures/dist",
-    "test:e2e:serve": "npx serve -n -S -L -p 8274",
+    "test:e2e:serve": "npx serve -n -S -L -p 8274 --config ./tests/config/serve.json",
     "test:e2e:start": "npm run test:e2e:instrument && npm run test:e2e:serve",
     "cy:run": "cypress run",
     "cy:run:record": "cypress run --record",
